git-svn-id: svn://123.57.173.36/meijia-parent@430 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ meijia-utils/src/main/java/com/meijia/utils/ExcelUtil.java
@@ -1,10 +1,6 @@
 package com.meijia.utils;
 
 import java.awt.Font;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -13,13 +9,6 @@ import org.apache.poi.ss.usermodel.CellStyle;
 import org.apache.poi.ss.usermodel.IndexedColors;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
-import org.apache.poi.xssf.usermodel.XSSFCell;
-import org.apache.poi.xssf.usermodel.XSSFRichTextString;
-import org.apache.poi.xssf.usermodel.XSSFRow;
-import org.apache.poi.xssf.usermodel.XSSFSheet;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
-
-import com.sun.rowset.internal.Row;
 
 
 
@@ -382,7 +371,7 @@ public class ExcelUtil {
         for (short i = 1; i < list.size(); i++) {
             // Row 行,Cell 方格 , Row 和 Cell 都是从0开始计数的
             // 创建一行，在页sheet上
-            org.apache.poi.ss.usermodel.Row row1 = sheet.createRow((short) i);
+            org.apache.poi.ss.usermodel.Row row1 = sheet.createRow(i);
             // 在row行上创建一个方格
             for(short j=0;j<keys.length;j++){
                 org.apache.poi.ss.usermodel.Cell cell = row1.createCell(j);
